half done

+++ includes/corsair.h
@@ -6,7 +6,7 @@
 /*   By: ahammoud <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/18 21:39:32 by ahammoud          #+#    #+#             */
-/*   Updated: 2022/07/22 23:11:01 by ahammoud         ###   ########.fr       */
+/*   Updated: 2022/07/23 16:37:05 by ahammoud         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,20 +27,20 @@
 #include <openssl/x509.h>
 
 typedef	struct s_var{
-	unsigned	long	m;
-	unsigned	long	q;
-	unsigned	long	p;
-	unsigned	long	pubexp;
-	unsigned	long	privexp;
-	unsigned	long	e1;
-	unsigned	long	e2;
-	unsigned	long	coeff;
-	unsigned	long	phi;
+	long	double	m;
+	long	double	q;
+	long	double	p;
+	long	double	pubexp;
+	long	double	privexp;
+	long	double	e1;
+	long	double	e2;
+	long	double	coeff;
+	long	double	phi;
 }	t_var;
 
 typedef struct s_prim{
-	unsigned	long	p;
-	unsigned	long	q;
+	long	double	p;
+	long	double	q;
 }	t_prime;
 
 
@@ -48,8 +48,8 @@ typedef struct s_prim{
 char				*get_data(char *str);
 char				*trim(char *buff);
 char				*get_data(char *str);
-unsigned	long	get_modulus(char *str);
-unsigned	long	ft_fnprime(unsigned	long	nb);
-unsigned	long	ft_primefactor(unsigned	long nb);
+long	double	get_modulus(char *str);
+long	double	ft_fnprime(long	double	nb);
+long	double	ft_primefactor(long	double nb);
 
 #endif
